+++ test/services/isochrones_tests.dart
@@ -8,10 +8,8 @@ void isochronesTests({
   test(
     'Fetch Isochrone Data using [getIsochrones]',
     () async {
-      final IsochroneData isochroneData = await service.getIsochrones(
-        locations: coordinates,
-        range: <int>[300, 200],
-      );
+      final GeoJsonFeatureCollection isochroneData = await service
+          .getIsochrones(locations: coordinates, range: <int>[300, 200]);
       expect(isochroneData.bbox.length, 2);
       expect(isochroneData.features.length, greaterThan(0));
       expect(isochroneData.features.first.properties.groupIndex, 0);